requirements.txt
flask
opencv-python
cmake
dlib
numpy
face-recognition
pymongo[srv]
